+++ lib/mir-fmm/src/function_declaration.rs
@@ -3,7 +3,7 @@ use crate::{context::Context, reference_count, type_};
 pub fn compile(context: &Context, declaration: &mir::ir::FunctionDeclaration) {
     context.module_builder().declare_variable(
         declaration.name(),
-        reference_count::compile_type_with_reference_count(type_::compile_unsized_closure(
+        reference_count::block::compile_type(type_::compile_unsized_closure(
             declaration.type_(),
             context.types(),
         )),
