+++ lib/mir-fmm/src/function_definition.rs
@@ -10,7 +10,7 @@ pub fn compile(
 ) -> Result<(), CompileError> {
     context.module_builder().define_variable(
         definition.name(),
-        reference_count::compile_static(closure::compile_content(
+        reference_count::block::compile_static(closure::compile_content(
             entry_function::compile(context, definition, true, global_variables)?,
             closure::metadata::compile(context, definition)?,
             fmm::ir::Undefined::new(type_::compile_closure_payload(definition, context.types())),
