+++ lib/mir-fmm/src/lib.rs
@@ -105,7 +105,7 @@ fn compile_global_variables(
                 declaration.name().into(),
                 fmm::build::variable(
                     declaration.name(),
-                    fmm::types::Pointer::new(reference_count::compile_type_with_reference_count(
+                    fmm::types::Pointer::new(reference_count::block::compile_type(
                         type_::compile_unsized_closure(declaration.type_(), types),
                     )),
                 ),
@@ -116,7 +116,7 @@ fn compile_global_variables(
                 declaration.name().into(),
                 fmm::build::variable(
                     declaration.name(),
-                    fmm::types::Pointer::new(reference_count::compile_type_with_reference_count(
+                    fmm::types::Pointer::new(reference_count::block::compile_type(
                         type_::compile_unsized_closure(declaration.type_(), types),
                     )),
                 ),
@@ -126,16 +126,14 @@ fn compile_global_variables(
             (
                 definition.name().into(),
                 fmm::build::bit_cast(
-                    fmm::types::Pointer::new(reference_count::compile_type_with_reference_count(
+                    fmm::types::Pointer::new(reference_count::block::compile_type(
                         type_::compile_unsized_closure(definition.type_(), types),
                     )),
                     fmm::build::variable(
                         definition.name(),
-                        fmm::types::Pointer::new(
-                            reference_count::compile_type_with_reference_count(
-                                type_::compile_sized_closure(definition, types),
-                            ),
-                        ),
+                        fmm::types::Pointer::new(reference_count::block::compile_type(
+                            type_::compile_sized_closure(definition, types),
+                        )),
                     ),
                 )
                 .into(),
