+++ lib/mir-fmm/src/reference_count.rs
@@ -1,3 +1,4 @@
+pub mod block;
 mod count;
 mod expression;
 mod function;
@@ -6,15 +7,4 @@ pub mod pointer;
 pub mod record;
 pub mod variant;
 
-use crate::CompileError;
 pub use expression::*;
-
-pub fn compile_static(
-    expression: impl Into<fmm::build::TypedExpression>,
-) -> Result<fmm::build::TypedExpression, CompileError> {
-    Ok(fmm::build::record(vec![count::compile_static()?, expression.into()]).into())
-}
-
-pub fn compile_type_with_reference_count(type_: impl Into<fmm::types::Type>) -> fmm::types::Record {
-    fmm::types::Record::new(vec![count::compile_type().into(), type_.into()])
-}
